Update to the Readme-edits branch
I wrote a small description of what I am doing to learn how to use Github and why I am learning it so that in the future if all goes well and I actually learn a little bit of how to use Ruby in a Windows environment without the use of a virtual machine I can then use the Git to collaborate and edit projects and research I am doing for school with others. Github is collaboration it seems. Very cool.